suffix-tree_array/requirements.txt
# Python requirements for DNA Pattern Matching - Suffix Array Implementation
# Install with: pip install -r requirements.txt

# Optional: For FASTA parsing and NCBI downloads
biopython>=1.79

# Optional: For running tests with biopython>=1.81
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
pandas>=2.0
memory_profiler>=0.61
psutil>=5.9
pyyaml>=6.0
requests>=2.31
pytest>=7.4
pytest-cov>=4.1
tqdm>=4.66
pytest>=7.0.0

# Optional: For performance profiling
memory-profiler>=0.60.0

# Optional: For visualization (if you extend the project)
matplotlib>=3.5.0
seaborn>=0.11.0

# Optional: For scientific computing
numpy>=1.21.0
